requirements.txt
asgiref==3.8.1
Django==4.2.14
django-ckeditor==6.7.1
django-environ==0.11.2
django-filter==24.3
django-js-asset==2.2.0
django-query-counter==0.4.0
djangorestframework==3.15.2
Markdown==3.7
pillow==10.4.0
sqlparse==0.5.1
tabulate==0.9.0
tzdata==2024.1
